requirements.txt
sentence-transformers
transformers
 keybert
 openai
 pinecone
 python-dotenv
 PyPDF2